+++ pkg/oc/cli/cli.go
@@ -25,17 +25,17 @@ import (
 	"github.com/openshift/oc/pkg/cli/logs"
 	"github.com/openshift/oc/pkg/cli/observe"
 	"github.com/openshift/oc/pkg/cli/options"
+	"github.com/openshift/oc/pkg/cli/recycle"
 	"github.com/openshift/oc/pkg/cli/secrets"
 	"github.com/openshift/oc/pkg/cli/serviceaccounts"
 	"github.com/openshift/oc/pkg/cli/whoami"
 	"github.com/openshift/oc/pkg/helpers/term"
 	"github.com/openshift/origin/pkg/cmd/flagtypes"
 	"github.com/openshift/origin/pkg/cmd/infra/deployer"
-	"github.com/openshift/origin/pkg/cmd/recycle"
 	"github.com/openshift/origin/pkg/cmd/templates"
 	"github.com/openshift/origin/pkg/oc/cli/admin"
 	"github.com/openshift/origin/pkg/oc/cli/admin/buildchain"
-	sync "github.com/openshift/origin/pkg/oc/cli/admin/groups/sync"
+	"github.com/openshift/origin/pkg/oc/cli/admin/groups/sync"
 	"github.com/openshift/origin/pkg/oc/cli/buildlogs"
 	"github.com/openshift/origin/pkg/oc/cli/cancelbuild"
 	"github.com/openshift/origin/pkg/oc/cli/debug"
