+++ staging/src/github.com/openshift/oc/pkg/cli/recycle/recycle_test.go
