+++ staging/src/github.com/openshift/oc/pkg/cli/recycle/walker.go
