+++ staging/src/github.com/openshift/oc/pkg/cli/recycle/walker_other.go
